feat(xsd): added test
