+++ packages/cfdi/xsd/config/index.js
@@ -1,0 +1,2 @@
+'use strict'
+module.exports = require('./xsd.cjs.production.min.js')
