+++ packages/cfdi/xsd/package.index.json
@@ -1,0 +1,5 @@
+{
+  "main": "dist/index.js",
+  "typings": "dist/index.d.ts",
+  "module": "dist/xsd.esm.js"
+}
