+++ packages/cfdi/xsd/test/emisor.test.ts
@@ -1,28 +1,5 @@
 import { describe, expect, it, test } from 'vitest';
-
-import { Schema } from '../src'; // Reemplaza con la ruta correcta a tu clase Schema
-import path from 'path';
-
-// Configura la instancia de Schema
-const schema = Schema.of();
-const files = path.join(
-  path.resolve(
-    __dirname,
-    '..',
-    '..',
-    '..',
-    'cfdi',
-    'schema',
-    'src',
-    'files',
-    'schema'
-  )
-);
-
-schema.setConfig({
-  path: files,
-  
-});
+import { schema } from './config';
 
 const emisor = schema.cfdi.emisor;
 describe('Emisor', () => {
