+++ packages/cfdi/xsd/test/relacionado.test.ts
@@ -1,0 +1,7 @@
+import { describe, expect, it, test } from 'vitest';
+import { schema } from './config';
+
+const emisor = schema.cfdi.emisor;
+describe('Relacionado', () => { 
+    it('Debería validar un objeto de Relacionado válido', () => {})
+})
